third_party/blink/web_tests/platform/win11-arm64/http/tests/eventsource/eventsource-reconnect-on-redirect-expected.txt

CONSOLE ERROR: EventSource's response has a MIME type ("text/html") that is not "text/event-stream". Aborting the connection.
Test EventSource reconnect after end of event stream.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS es.readyState is es.CONNECTING
FAIL es.readyState should be 0. Was 2.
FAIL reconnect timed out
PASS successfullyParsed is true

TEST COMPLETE
